chore: removes unused files and implements mathjs
