+++ app/index.html
@@ -61,7 +61,7 @@
 				<option value="10">10</option>
 
 			</select>
-			
+
 		</div>
 	</div>
 
@@ -92,7 +92,8 @@
 
 	</section>
 
-	<script src="script.js"></script>
+    <script async src="calculator.js"></script>
+    <script defer src="dom.js"></script>
 </body>
 
 </html>
