+++ constants/operators-as-text.js
@@ -1,0 +1,16 @@
+const operatorsAsText = {
+	plus: '+',
+	'added to': '+',
+	adds: '+',
+	with: '+',
+	minus: '-',
+	subtract: '-',
+	less: '-',
+	'divided by': '/',
+	by: '/',
+	'multiplied by': '*',
+	into: '*',
+	cross: '*'
+};
+
+module.exports = operatorsAsText;
