+++ deprecated/coreCalc.js
