+++ deprecated/coreCalc.test.js
