+++ deprecated/coreConv.js
