+++ deprecated/coreConv.test.js
