+++ deprecated/defaultRates.js
